requirements.txt
annotated-types==0.6.0
data-traps-core @ file:///misc/jb/data-traps-core
importlib-metadata==7.0.0
pydantic==2.5.2
pydantic-yaml==1.2.0
pydantic_core==2.14.5
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
setuptools==68.2.2
typing_extensions==4.9.0
wheel==0.41.2
zipp==3.17.0
